+++ src/Basescape/ManufactureState.cpp
@@ -265,7 +265,7 @@ void ManufactureState::lstManufactureClickMiddle(Action *)
  */
 void ManufactureState::lstManufactureMousePress(Action *action)
 {
-	if (!_lstManufacture->isInsideNoScrollArea(action->getAbsoluteXMouse()))
+	if (!_lstManufacture->isInsideNoScrollArea(static_cast<int>(action->getAbsoluteXMouse())))
 	{
 		return;
 	}
